Worked on getting the command working Done:

+++ src/main/java/com/gabezter/MoblieChest/Listen.java
@@ -6,7 +6,12 @@ import java.util.List;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
+import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.World;
+import org.bukkit.block.Block;
+import org.bukkit.block.BlockFace;
+import org.bukkit.block.Chest;
 import org.bukkit.block.Sign;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
@@ -16,10 +21,12 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.event.block.SignChangeEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.inventory.Inventory;
 
 public class Listen implements Listener {
 
 	Main plugin;
+	Methods method;
 
 	public Listen(Main plugin) {
 		this.plugin = plugin;
@@ -51,12 +58,6 @@ public class Listen implements Listener {
 
 	@EventHandler
 	public void signRead(SignChangeEvent e) {
-		int x = e.getBlock().getX();
-		int y = e.getBlock().getY();
-		int z = e.getBlock().getZ();
-		String sx = Integer.toString(x);
-		String sy = Integer.toString(y);
-		String sz = Integer.toString(z);
 
 		plugin.userFile = new File(plugin.users, e.getPlayer().getUniqueId().toString() + ".yml");
 		File user = plugin.userFile;
@@ -64,14 +65,29 @@ public class Listen implements Listener {
 
 		FileConfiguration userConfig = plugin.userconfig;
 
-		if (e.getLine(0).equalsIgnoreCase("Chest Access")
+		if ((getAttachedBlock(e.getBlock()) != null)
+		        && e.getLine(0).equalsIgnoreCase("Chest Access")
 		        && e.getLine(1).equalsIgnoreCase(e.getPlayer().getName())
 		        && !(e.getLine(2).equals(""))) {
+
+			int x = e.getBlock().getRelative(getAttachedBlock(e.getBlock())).getX();
+			int y = e.getBlock().getRelative(getAttachedBlock(e.getBlock())).getY();
+			int z = e.getBlock().getRelative(getAttachedBlock(e.getBlock())).getZ();
+			String sx = Integer.toString(x);
+			String sy = Integer.toString(y);
+			String sz = Integer.toString(z);
 			String name = e.getLine(2);
-			plugin.chestFile = new File(plugin.chests, sx + "`" + sy + "`" + sz + "`" + e.getBlock().getWorld().getName() + "`" + name + ".yml");
+			String fileName = sx + "`" + sy + "`" + sz + "`" + e.getBlock().getWorld().getName() + "`" + name;
+
+			e.getPlayer().sendMessage(fileName);
+
+			plugin.chestFile = new File(plugin.chests, fileName + ".yml");
 			File chestConfig = plugin.chestFile;
 			plugin.chestconfig = YamlConfiguration.loadConfiguration(chestConfig);
 			FileConfiguration chest = plugin.chestconfig;
+			File chests1 = plugin.chestsFile;
+			FileConfiguration chests2 = plugin.chestsConfig;
+
 			chest.set("Name", name);
 			chest.set("X", x);
 			chest.set("Y", y);
@@ -79,17 +95,19 @@ public class Listen implements Listener {
 			chest.set("World", e.getBlock().getWorld());
 			chest.set("Users", e.getPlayer().getUniqueId().toString());
 			chest.set("Users." + e.getPlayer().getUniqueId().toString(), e.getPlayer().getName());
-			try {
-				chest.save(chestConfig);
-			}
-			catch (IOException e2) {
-			}
 
 			List<String> chests = userConfig.getStringList("Chests");
-			chests.add(sx + "`" + sy + "`" + sz + "`" + name);
+			chests.add(fileName);
 			userConfig.set("Chests", chests);
+
+			List<String> chestes = chests2.getStringList("Chests");
+			chestes.add(fileName);
+			chests2.set("Chests", chestes);
+
 			try {
+				chests2.save(chests1);
 				userConfig.save(user);
+				chest.save(chestConfig);
 			}
 			catch (IOException e1) {
 				Bukkit.getServer()
@@ -115,14 +133,13 @@ public class Listen implements Listener {
 		String sy = Integer.toString(y);
 		String sz = Integer.toString(z);
 		e.getPlayer().sendMessage(e.getBlock().toString());
-		if (e.getBlock().getType().equals(Material.SIGN_POST)) {
-			Sign sign = (Sign) e.getBlock();
+		if (e.getBlock().getType().equals(Material.SIGN_POST) || e.getBlock().getType().equals(Material.WALL_SIGN) || e.getBlock().getType().equals(Material.SIGN)) {
+			Sign sign = (Sign) e.getBlock().getState();
 			e.getPlayer().sendMessage(sign.getLine(1));
 			if (sign.getLine(1).equals(e.getPlayer().getName())) {
 				plugin.userFile = new File(plugin.users, e.getPlayer().getUniqueId().toString() + ".yml");
 				File user = plugin.userFile;
 				plugin.userconfig = YamlConfiguration.loadConfiguration(user);
-
 				String name = sign.getLine(2);
 				FileConfiguration userConfig = plugin.userconfig;
 				plugin.chestFile = new File(plugin.chests, sx + "`" + sy + "`" + sz + "`" + e.getBlock().getWorld().getName() + "`" + name + ".yml");
@@ -130,16 +147,27 @@ public class Listen implements Listener {
 				plugin.chestconfig = YamlConfiguration.loadConfiguration(chestConfig);
 				FileConfiguration chest = plugin.chestconfig;
 
+				File chests1 = plugin.chestsFile;
+				FileConfiguration chests2 = plugin.chestsConfig;
+
 				userConfig.getStringList("Chests").remove(sx + "`" + sy + "`" + sz + "`" + e.getBlock().getWorld().getName() + "`" + name);
-				chestConfig.delete();
+				chests2.getStringList("Chests").remove(sx + "`" + sy + "`" + sz + "`" + e.getBlock().getWorld().getName() + "`" + name);
+
+				if (chestConfig.exists()) {
+					chestConfig.delete();
+				}
+				else {
+					Bukkit.getServer().getLogger().log(null, "Config file tried to be deleted but it didn't exist in the first place!");
+					Bukkit.getServer().getLogger().log(null, "Config file name: " + sx + "`" + sy + "`" + sz + "`" + e.getBlock().getWorld().getName() + "`" + name);
+				}
 				try {
+					chests2.save(chests1);
 					userConfig.save(user);
 					chest.save(chestConfig);
 				}
 				catch (IOException e1) {
 				}
-				e.getPlayer().sendMessage(
-				        ChatColor.DARK_GREEN + "Chest unlinked Successful!!");
+				e.getPlayer().sendMessage(ChatColor.DARK_GREEN + "Chest unlinked Successful!!");
 				e.setCancelled(false);
 			}
 			e.getPlayer().sendMessage(ChatColor.DARK_RED + "You can't break this Sign!!!");
@@ -148,8 +176,45 @@ public class Listen implements Listener {
 		e.setCancelled(false);
 	}
 
-	public String chestName(int x, int y, int z, String sworld, Player player) {
-
+	public String getChest(Player player, String string) {
+		String uuid = player.getUniqueId().toString();
+		
+		plugin.userFile = new File(plugin.users, uuid + ".yml");
+		File user = plugin.userFile;
+		plugin.userconfig = YamlConfiguration.loadConfiguration(user);
+		FileConfiguration userConfig = plugin.userconfig;
+		
+		List<String> chests = userConfig.getStringList("Chests");
+		for(String i : chests){
+			if(i.contains(string)){
+				return i;
+			}
+		}
 		return null;
 	}
+
+	public BlockFace getAttachedBlock(Block sb) {
+		BlockFace[] blockFaces = { BlockFace.EAST, BlockFace.NORTH, BlockFace.WEST, BlockFace.SOUTH };
+		for (BlockFace bf : blockFaces) {
+			Block bu = sb.getRelative(bf);
+			if ((bu.getType().equals(Material.CHEST))) {
+				return bf;
+			}
+		}
+		return null;
+	}
+
+	public Inventory getChestInventory(int x, int y, int z, World world){
+		Inventory inv = null;
+		Location loc = new Location(world, x, y, z);
+		if(loc.getBlock().getType().equals(Material.CHEST)){
+			Block block = loc.getBlock();
+			Chest chest = (Chest) block.getState();
+			inv = chest.getBlockInventory();
+			return inv;
+		}
+		
+		return inv;
+	}
+
 }
